requirements.txt
flask==3.0.2
python-dotenv==1.0.1
flask-cors==4.0.0 
supabase==2.10.0
boto3==1.34.69
requests==2.32.3
openai>=1.0.0